aider: Changed the version of the `nexus-staging-maven-plugin` to `1.6.7`.
# Context:
USER: That did not fix the error. Now getting """[ERROR] Unresolveable build extension: Plugin org.sonatype.plugins:nexus-staging-maven-plugin:1.6.8 or one of its dependencies could not be resolved: Failed to read artifact descriptor for org.sonatype.plugins:nexus-staging-maven-plugin:jar:1.6.8 @"""
ASSISTANT: The error message indicates that Maven is unable to resolve the `nexus-staging-maven-plugin` dependency. This could be due to several reasons:

1. The plugin version is incorrect or not available in the repositories specified.
2. There is a network issue preventing Maven from accessing the repositories.
3. There is a typo or configuration error in your `pom.xml` file.

To resolve this issue, you can try the following steps:

1. Check if the `nexus-staging-maven-plugin` version `1.6.8` is available in your repositories. If it's not, you need to change it to a version that is available.
2. Check your network connection and make sure your repositories are accessible.
3. Check your `pom.xml` file for any configuration errors.

If the plugin version is not available in your repositories and you cannot change it, you can try using a different version of the plugin that is available. For example, you can try using version `1.6.7` instead of `1.6.8`.

Here is the change:

```xml
pom.xml
<<<<<<< ORIGINAL
      <plugin>
        <groupId>org.sonatype.plugins</groupId>
        <artifactId>nexus-staging-maven-plugin</artifactId>
        <version>1.6.8</version>
        <extensions>true</extensions>
        <configuration>
          <serverId>ossrh</serverId>
          <nexusUrl>https://oss.sonatype.org/</nexusUrl>
          <autoReleaseAfterClose>true</autoReleaseAfterClose>
        </configuration>
      </plugin>
=======
      <plugin>
        <groupId>org.sonatype.plugins</groupId>
        <artifactId>nexus-staging-maven-plugin</artifactId>
        <version>1.6.7</version>
        <extensions>true</extensions>
        <configuration>
          <serverId>ossrh</serverId>
          <nexusUrl>https://oss.sonatype.org/</nexusUrl>
          <autoReleaseAfterClose>true</autoReleaseAfterClose>
        </configuration>
      </plugin>
>>>>>>> UPDATED
```

This will change the version of the `nexus-staging-maven-plugin` to `1.6.7`. Please try this and let me know if it resolves the issue.