aider: Changed the goal of the `nexus-maven-plugin` to `staging-release`.
# Context:
USER: getting """[ERROR] Could not find goal 'validate' in plugin org.sonatype.plugins:nexus-maven-plugin:1.9.2.4 among available goals staging-profiles-list, staging-drop, staging-release, staging-list, staging-build-promotion, settings-download, staging-close, staging-promote """ on mvn deploy
ASSISTANT: The error message indicates that the `validate` goal is not available in the `nexus-maven-plugin` version `1.9.2.4`.

The `nexus-maven-plugin` does not have a `validate` goal. The available goals for `nexus-maven-plugin` are `staging-profiles-list`, `staging-drop`, `staging-release`, `staging-list`, `staging-build-promotion`, `settings-download`, `staging-close`, and `staging-promote`.

If you want to validate the project before deploying, you can use the `maven-validate-plugin` or the `validate` phase of the Maven lifecycle.

If you want to use the `nexus-maven-plugin`, you need to choose one of the available goals. For example, if you want to release the staging repository, you can use the `staging-release` goal.

Here is the change:

```xml
pom.xml
<<<<<<< ORIGINAL
      <plugin>
        <groupId>org.sonatype.plugins</groupId>
        <artifactId>nexus-maven-plugin</artifactId>
        <version>1.9.2.4</version>
        <executions>
          <execution>
            <id>default-cli</id>
            <goals>
              <goal>validate</goal>
            </goals>
          </execution>
        </executions>
      </plugin>
=======
      <plugin>
        <groupId>org.sonatype.plugins</groupId>
        <artifactId>nexus-maven-plugin</artifactId>
        <version>1.9.2.4</version>
        <executions>
          <execution>
            <id>default-cli</id>
            <goals>
              <goal>staging-release</goal>
            </goals>
          </execution>
        </executions>
      </plugin>
>>>>>>> UPDATED
```

This will change the goal of the `nexus-maven-plugin` to `staging-release`. Please try this and let me know if it resolves the issue.