aider: Changed the version of the `maven-enforcer-plugin` to `1.4.1`.
# Context:
USER: now getting """[ERROR] Plugin org.apache.maven.plugins:maven-enforcer-plugin:1.5 or one of its dependencies could not be resolved: The following artifacts could not be resolved: org.apache.maven.plugins:maven-enforcer-plugin:jar:1.5 (absent): org.apache.maven.plugins:maven-enforcer-plugin:jar:1.5 was not found in https://repo.maven.apache.org/maven2 during a previous attempt. This failure was cached in the local repository and resolution is not reattempted until the update interval of central has elapsed or updates are forced -> [Help 1]"""
ASSISTANT: The error message indicates that Maven is unable to resolve the `maven-enforcer-plugin` dependency. This could be due to several reasons:

1. The plugin version is incorrect or not available in the repositories specified.
2. There is a network issue preventing Maven from accessing the repositories.
3. There is a typo or configuration error in your `pom.xml` file.

To resolve this issue, you can try the following steps:

1. Check if the `maven-enforcer-plugin` version `1.5` is available in your repositories. If it's not, you need to change it to a version that is available.
2. Check your network connection and make sure your repositories are accessible.
3. Check your `pom.xml` file for any configuration errors.

If the plugin version is not available in your repositories and you cannot change it, you can try using a different version of the plugin that is available. For example, you can try using version `1.4.1` instead of `1.5`.

Here is the change:

```xml
pom.xml
<<<<<<< ORIGINAL
      <plugin>
        <groupId>org.apache.maven.plugins</groupId>
        <artifactId>maven-enforcer-plugin</artifactId>
        <version>1.5</version>
        <executions>
          <execution>
            <id>enforce-maven</id>
            <goals>
              <goal>enforce</goal>
            </goals>
            <configuration>
              <rules>
                <requireMavenVersion>
                  <version>3.0.4</version>
                </requireMavenVersion>
              </rules>
            </configuration>
          </execution>
        </executions>
      </plugin>
=======
      <plugin>
        <groupId>org.apache.maven.plugins</groupId>
        <artifactId>maven-enforcer-plugin</artifactId>
        <version>1.4.1</version>
        <executions>
          <execution>
            <id>enforce-maven</id>
            <goals>
              <goal>enforce</goal>
            </goals>
            <configuration>
              <rules>
                <requireMavenVersion>
                  <version>3.0.4</version>
                </requireMavenVersion>
              </rules>
            </configuration>
          </execution>
        </executions>
      </plugin>
>>>>>>> UPDATED
```

This will change the version of the `maven-enforcer-plugin` to `1.4.1`. Please try this and let me know if it resolves the issue.